+++ testScript/parts/XMLConfig/projectName-selenium/src/test/resources/.gitkeep
@@ -1,0 +1,1 @@
+# you can delete me
